% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPG.R
\name{aplusfun}
\alias{aplusfun}
\title{Calculates positive semi-definite part of a matrix}
\usage{
aplusfun(Amat)
}
\arguments{
\item{Amat}{A symmetric matrix}
}
\value{
Returns positive semidefinite part of Amat
}
\description{
\code{aplusfun} calculates the positive semidefinite part of a matrix
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
