% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPGtest.R
\name{calculateLRstat}
\alias{calculateLRstat}
\title{\code{calculateLRstat} calculates the likelihood ratio statistic}
\usage{
calculateLRstat(A, null, alt)
}
\arguments{
\item{A}{List of symmetric A matrices}

\item{null}{Model fitted under the null}

\item{alt}{Model fitted under the alternative}
}
\value{
Value of the likelihood ratio statistic
}
\description{
\code{calculateLRstat} calculates the likelihood ratio statistic
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
