% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPG.R
\name{calculateobjfun}
\alias{calculateobjfun}
\title{\code{calculateobjfun} calculates the objective function}
\usage{
calculateobjfun(Aplus, Lambda, U)
}
\arguments{
\item{Aplus}{List of positive semi definite part of A matrices}

\item{Lambda}{List of Lambdas}

\item{U}{U matrix that contains the othogonal vectors}
}
\value{
Returns the value of the objective function, sum(||Aplus-U Lambda U^T||_F^2)
}
\description{
\code{calculateobjfun} calculates the objective function
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
