% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.multiRDPGfit}
\alias{plot.multiRDPGfit}
\title{Plots object from \code{multiRDPG}}
\usage{
\method{plot}{multiRDPGfit}(x, ...)
}
\arguments{
\item{x}{multiRDPGfit object from function \code{multiRDPG}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plots object from \code{multiRDPG}
}
\seealso{
\code{\link{multiRDPG}}
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
