% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot.multiRDPGtest}
\alias{plot.multiRDPGtest}
\title{Plots object from \code{multiRDPG_test}}
\usage{
\method{plot}{multiRDPGtest}(x, ...)
}
\arguments{
\item{x}{multiRDPGtest object from function \code{multiRDPG_test}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plots histogram of permutation test statistics and indicates test statistic value with red line.
}
\details{
Red line indicates the value of the test statistics with a red line.
}
\seealso{
\code{\link{multiRDPG_test}}
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
