% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.multiRDPGtest}
\alias{print.multiRDPGtest}
\title{Print object from \code{multiRDPG_test}}
\usage{
\method{print}{multiRDPGtest}(x, ...)
}
\arguments{
\item{x}{multiRDPGtest object from function \code{multiRDPG_test}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print object from \code{multiRDPG_test}
}
\seealso{
\code{\link{multiRDPG_test}}
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
