% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPGtest.R
\name{swapA}
\alias{swapA}
\title{\code{swapA} swaps the elements of A at random}
\usage{
swapA(A)
}
\arguments{
\item{A}{List of symmetric A matrices}
}
\value{
Returns list of matrices with elements swapped
}
\description{
\code{swapA} swaps the elements of A at random
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
