% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPG.R
\name{updateL}
\alias{updateL}
\title{\code{updateL} updates all Lambda}
\usage{
updateL(Aplus, U, d)
}
\arguments{
\item{Aplus}{List of positive definite parts of A}

\item{U}{Current U matrix}

\item{d}{Dimension of latent space}
}
\value{
Returns list of updated Lambdas
}
\description{
\code{updateL} updates all Lambda
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
