% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRDPG.R
\name{updateU}
\alias{updateU}
\title{\code{updateU} calculates the U update}
\usage{
updateU(AA, Lambda, Uold)
}
\arguments{
\item{AA}{All Aplus matrices column binded}

\item{Lambda}{List of Lambdas}

\item{Uold}{U found in past iteration}
}
\value{
Returns updated U
}
\description{
\code{updateU} calculates the U update
}
\author{
Agnes Martine Nielsen (agni@dtu.dk)
}
\keyword{internal}
